regular fragmentationupdated with input checks, perfectly working

+++ runner.py
@@ -6,36 +6,45 @@ from FirewallHandler import FirewallHandler
 from log import log
 from input_handler import inputHandler
 from frammentizzatore import frammentizzatore
-from scapy.all import send, defragment6, IPv6, raw
+from scapy.all import send
 
 
 logs_handler = None
-
+input_handler = None
 
 def sendFragments(fragments):
+    num_of_fragments = len(fragments)
+    logs_handler.logger.info("Sending %d fragments", num_of_fragments)
+    for frag in fragments:
+        frag.show()
     send(fragments)
+    logs_handler.logger.info("Fragments sent")
 
 def traffic_handler(packet):
     
-    logs_handler.logger.debug("Traffic intercepted")
-    framm = frammentizzatore()
-    fragments = framm.fragment(packet)
+    logs_handler.logger.info("Traffic intercepted")
+    framm = frammentizzatore(logs_handler)
+    fragments = framm.fragment(packet, input_handler.fragmentSize)
+    
+    if fragments == None:
+        packet.accept()
+        return
+    
     #packet.set_payload(bytes(fragments))
-    #packet.accept()
     #fragments.show()
     sendFragments(fragments)
     #p = defragment6(fragments)
     #packet.accept()
     packet.drop()
-    
     return
 
 
-def setFirewallRules(protocol = "icmpv6", dest_ipv6addr="", dstPort = ""):
+def setFirewallRules(logs_handler, protocol = "", dest_ipv6addr="", dstPort = ""):
     
     os_type = platform.system().lower()
 
-    firewall_handler = FirewallHandler(os_type, protocol, dest_ipv6addr, dstPort)
+    logs_handler.logger.info("Inserting firewall rules...")
+    firewall_handler = FirewallHandler(os_type, logs_handler, protocol, dest_ipv6addr, dstPort)
     firewall_handler.insert_firewall_rules()
 
     return firewall_handler
@@ -51,13 +60,14 @@ def main():
         logs_handler.logger.error("input.json not found")
         exit(1)
 
-    input_handler = inputHandler(input_file)
+    global input_handler
+    input_handler = inputHandler(input_file, logs_handler)
     
     res = input_handler.parse_input()
     if res == False:
         exit(1)
     
-    firewall_handler = setFirewallRules()
+    firewall_handler = setFirewallRules(logs_handler, input_handler.protocol, input_handler.ipv6Dest, input_handler.dstPort)
     logs_handler.logger.info("Firewall rules set")
 
     nfqueue = NetfilterQueue()
@@ -69,10 +79,12 @@ def main():
     except KeyboardInterrupt:
         logs_handler.logger.info("Frammentizzatore Stopped")
     finally:
+        logs_handler.logger.info("Deleting firewall rules")
         firewall_handler.delete_firewall_rules()
+        logs_handler.logger.info("Firewall rules deleted")
         nfqueue.unbind()
     
-    print("Exiting...")
+    logs_handler.logger.info("Exiting...")
